Move StateTransitionException code into job manager
